+++ xeo/xeo.py
@@ -1,4 +1,5 @@
 from .axioms import Catalogue, Instrument, Instruments, catalogue, instruments
+from .plot import plot_bands, plot_srf
 
 __all__ = [
     "Catalogue",
@@ -6,4 +7,6 @@ __all__ = [
     "Instruments",
     "catalogue",
     "instruments",
+    "plot_bands",
+    "plot_srf",
 ]